+++ lizmap/dialogs/html_editor.py
@@ -7,6 +7,7 @@ import logging
 from qgis.core import QgsVectorLayer
 from qgis.PyQt.QtWidgets import QDialog, QDialogButtonBox, QVBoxLayout
 
+from lizmap.qgis_plugin_tools.tools.i18n import tr
 from lizmap.widgets.html_editor import HtmlEditorWidget
 
 LOGGER = logging.getLogger('Lizmap')
@@ -14,10 +15,11 @@ LOGGER = logging.getLogger('Lizmap')
 
 class HtmlEditorDialog(QDialog):
 
-    def __init__(self, layer: QgsVectorLayer = None):
+    def __init__(self):
         # noinspection PyArgumentList
         QDialog.__init__(self)
-        self.editor = HtmlEditorWidget(layer)
+
+        self.editor = HtmlEditorWidget(self)
 
         layout = QVBoxLayout()
         # noinspection PyArgumentList
@@ -34,3 +36,7 @@ class HtmlEditorDialog(QDialog):
         accept_button.clicked.connect(self.accept)
         cancel_button = self.button_box.button(QDialogButtonBox.Cancel)
         cancel_button.clicked.connect(self.reject)
+
+    def set_layer(self, layer: QgsVectorLayer):
+        self.setWindowTitle(tr("HTML maptip for the layer '{}'").format(layer.name()))
+        self.editor.set_layer(layer)