+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 * Improve the editing form with non-spatial layers
 * Improve the new layout panel for Lizmap Web Client 3.7
-* Add new HTML wizard for the QGIS HTML Maptip popup
+* Add a new HTML wizard for the
+  * QGIS HTML Maptip popup
+  * HTML template in the dataviz
+  * HTML layout in the dataviz
+  * plot description in the dataviz
 
 ## 3.10.1 - 2023-02-07